Rename EC_Logger to Logger, and mark EC_Logger exports as deprecated.
